Update processing based on Eze's feedback
rom the start of the project, we have worked with a report that has Early onset and syndromic epilepsy.
This is why we initially had snomed since it has the codes. Now we know that the lab uses PanelApp, we now have to adjust the process to accommodate other reports.

RedCap doesn’t have PanelApp integration. I have just finished a meeting with Rosie on the best approach to handle the PanelApp usage on the test title due to the nature of data extraction from RedCap. Hence, we would be using PanelApp moving forward.

On the third point, I have been relying on the chat with the lab. However, looking up the format, it’s as Stef specified. I would suggest we have it as specified - NM_001350626.1:c.1550G>T. I have updated Rosie on that too.